z_sync/requirements.txt
termcolor
MacFSEvents
PyYAML
ntfy
